backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
langchain-groq==0.2.1
langchain-core==0.3.30
pydantic==2.10.6
